🚚 Split console module into several submodules
