+++ snakeshell/console/__init__.py
@@ -1,0 +1,5 @@
+from snakeshell.console.stream import write, writeline, error
+from snakeshell.console.prompt import prompt
+
+
+__all__ = ['write', 'writeline', 'error', 'prompt']
